pyproject.toml
[project]
name = "p-v-s"
version = "0.1.0"
description = "Utility to compare different (AI) news providers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "exa-py>=2.3.0",
    "linkup-sdk>=0.2.8",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tavily-python>=0.7.21",
]
